Move globalize-specific utils to separate module.
Create a new module that contains globalize-specific utility functions
reducing duplication across the different formatter modules. Also move
`src/util` to `src/util/main`.
